chore(users): rename files
